refactor: rename dock layout files
