+++ packages/picasso.js/src/core/layout/dock/config.js
