+++ packages/picasso.js/src/core/layout/dock/settings-resolver.js
